introduce PublicFollowupMessage and EphemeralFollowupMessage
