+++ core/src/main/kotlin/behavior/interaction/PublicInteractionResponseBehavior.kt
@@ -1,11 +1,14 @@
-package dev.kord.core.behavior.channel
+package behavior.interaction
 
 import dev.kord.common.annotation.KordPreview
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.Kord
-import dev.kord.core.behavior.InteractionResponseBehavior
+import dev.kord.core.behavior.interaction.InteractionResponseBehavior
+import dev.kord.core.cache.data.MessageData
 import dev.kord.core.cache.data.toData
 import dev.kord.core.entity.Message
+import dev.kord.core.entity.interaction.PublicFollowupMessage
+import dev.kord.rest.builder.interaction.PublicFollowupMessageCreateBuilder
 import dev.kord.rest.builder.interaction.PublicInteractionResponseModifyBuilder
 import kotlin.contracts.ExperimentalContracts
 import kotlin.contracts.InvocationKind
@@ -37,8 +40,19 @@ fun PublicInteractionResponseBehavior(applicationId: Snowflake, token: String, k
 @KordPreview
 @OptIn(ExperimentalContracts::class)
 suspend inline fun PublicInteractionResponseBehavior.edit(builder: PublicInteractionResponseModifyBuilder.() -> Unit): Message {
-    contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE)  }
+    contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
     val request = PublicInteractionResponseModifyBuilder().apply(builder).toRequest()
     val response = kord.rest.interaction.modifyInteractionResponse(applicationId, token, request)
     return Message(response.toData(), kord)
+}
+
+
+@KordPreview
+@OptIn(ExperimentalContracts::class)
+suspend inline fun PublicInteractionResponseBehavior.followUp(builder: PublicFollowupMessageCreateBuilder.() -> Unit): PublicFollowupMessage {
+    contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
+    val request = PublicFollowupMessageCreateBuilder().apply(builder).toRequest()
+    val response = kord.rest.interaction.createFollowupMessage(applicationId, token, request)
+    val data = MessageData.from(response)
+    return PublicFollowupMessage(Message(data, kord), applicationId, token, kord)
 }
